+++ SonsSdk/ESonsScene.cs
@@ -1,0 +1,8 @@
+namespace SonsSdk;
+
+public enum ESonsScene
+{
+    Title,
+    Loading,
+    Game
+}